Enhance Mesh Statistics Types and Update MeshStatsView Component
- Added new properties to CurrentStats and introduced MonthlyDownload, PackageData, FilteredStats, and MeshStatsViewProps interfaces in types.ts for improved data handling.
- Refactored MeshStatsView component to utilize the new types, enhancing type safety and clarity in data representation.
- Updated the logic for displaying package downloads to reflect the new core package downloads metric over the last 12 months.
